Added: Responsiveness for HomePage and refactored product page
