+++ package.json
@@ -23,6 +23,7 @@
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-redux": "^7.2.6",
+    "react-responsive-carousel": "^3.2.22",
     "react-router-breadcrumbs-hoc": "^4.1.0",
     "react-router-dom": "^5.2.0",
     "react-scripts": "5.0.0",
