+++ public/index.html
@@ -24,6 +24,19 @@
       work correctly both with client-side routing and a non-root public URL.
       Learn how to configure a non-root public URL by running `npm run build`.
     -->
+    <style>
+      * {
+        box-sizing: border-box;
+        font-family: Roboto, Arial, sans-serif;
+        text-decoration: none;
+      }
+      body {
+        overflow-x: hidden;
+      }
+      /* a {
+        text-decoration: none;
+      } */
+    </style>
     <title>Online Shopping for Women, Men, Kids Fashion & Lifestyle - Myntra</title>
   </head>
   <body>
