+++ src/App.css
@@ -1,7 +1,7 @@
-/* .App {
+.App {
     text-align: left;
   }
-   */
+  
   .App-logo {
     height: 40vmin;
     pointer-events: none;
