+++ src/components/dashboard/footwear/footwear.css
@@ -25,5 +25,11 @@
 @media screen and (max-width:568px) {
    .footwearHeading{
        font-size: 1rem;
-   }    
+   } 
+   .footwear {
+   max-width: inherit;
+    overflow-x: hidden;
+    display: flex;
+    margin: auto;
+   }   
 }
