+++ src/components/dashboard/imageSlider/imageSlider.css
@@ -1,72 +1,22 @@
-.slider {
-  margin-top: 0.5rem;
-  margin-bottom: 0.5rem;
-  background-color: #fff;
-  margin-top: 2rem;
+.carousel-wrapper{
+  width:100%;
+  height:350px;
+  padding:10px;
+  padding-top: 2em;
 }
-
-.wrap {
-  position: relative;
-  cursor: pointer;
-  padding: 0.3rem;
-  border-radius: 0.2rem;
+.carousel-image{
+  width:100%;
+  height:350px
 }
-.wrap:hover {
-  border: 0.5px solid #f41cb2;
-  border-radius: 0.2rem;
-  transition: all 0.3s ease;
-}
-.wrap a {
-  position: relative;
-  display: block;
-  text-decoration: none;
-  border-radius: 0.2rem;
-}
-
-.wrap a img {
-  width: 100%;
-  height: 100%;
-  object-fit: contain;
-}
-.slick-prev {
-  opacity: 0 ;
-  width: 5vw ;
-  height: 100% ;
-  z-index: 1 ;
-}
-.slick-prev::before {
-  color: #000 ;
-}
-.slick-prev:hover {
-  opacity: 1 ;
-  transition: opacity 0.3s ease ;
-}
-
-.slick-next {
-  opacity: 0 ;
-  width: 5vw ;
-  height: 100% ;
-  z-index: 1 ;
-}
-.slick-next::before {
-  color: #000 ;
-}
-
-.slick-next:hover {
-  opacity: 1 ;
-  transition: opacity 0.3s ease;
-}
-
-.slick-dots li button::before {
-  font-size: 0.5rem ;
-  color: gray ;
-}
-.slick-dots li.slick-active button::before {
-  color: gray ;
-}
-
-@media screen and (max-width: 568px) {
-  .wrap {
-    border: 1px solid goldenrod;
+@media only screen and (max-width: 600px) {
+  .carousel-image{
+      width: 100%;
+      height: 100%;
+      min-height: 184px;
+  }
+  .carousel-wrapper{
+      width: 100%;
+      height: 100%;
+      min-height: 184px;
   }
 }
