+++ src/components/dashboard/imageSlider/images.js
@@ -1,0 +1,15 @@
+/**
+ * @description Stores images for Carousel
+ * @author Sanjana Rao
+ */
+import slide1 from '../../../assets/image_1.jpg';
+import slide2 from '../../../assets/image_1.webp';
+import slide3 from '../../../assets/image_3.webp';
+import slide4 from '../../../assets/image_4.webp';
+import slide5 from '../../../assets/image_5.webp';
+import slide6 from '../../../assets/image_6.webp';
+import slide7 from '../../../assets/image_7.webp';
+
+const CarouselData = [slide1, slide2, slide3, slide4, slide5, slide6, slide7];
+
+export default CarouselData;
