+++ src/components/dashboard/indianWear/indianWear.css
@@ -50,3 +50,15 @@
     font-size: 1rem;
   }
 }
+
+@media screen and (max-width: 375px) {
+  .indianwearHeading {
+    font-size: 1rem;
+  }
+  .indianWear {
+    max-width: inherit;
+    overflow-x: hidden;
+    display: flex;
+    margin: auto;
+  }
+}
